pyproject.toml
[project]
name = "skos-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph==0.1.4",
    "ollama==0.1.3",  # pip install ollama (wrapper for the ollama HTTP API)
    "langchain-community==0.2.0",
    # "langchain-openai==0.1.0",
    "pypdf==4.0.2",
    "markdown2==2.4.0",
    "asciidoc",   # a tiny wrapper around the AsciiDoc parser
    "ebooklib==0.18",   # for EPUB
    "python-docx==1.1.2",
    "beautifulsoup4==4.12.3",
    "rdflib==7.0.0",
    "pypdf"
]
